+++ packages/CodingStandard/tests/Fixer/LineLength/BreakMethodArgumentsFixer/BreakMethodArgumentFixerTest.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\CodingStandard\Tests\Fixer\LineLength\BreakMethodArgumentFixer;
+namespace Symplify\CodingStandard\Tests\Fixer\LineLength\BreakMethodArgumentsFixer;
 
 use PhpCsFixer\Fixer\FixerInterface;
 use PhpCsFixer\WhitespacesFixerConfig;
-use Symplify\CodingStandard\Fixer\LineLength\BreakMethodArgumentFixer;
+use Symplify\CodingStandard\Fixer\LineLength\BreakMethodArgumentsFixer;
 use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 
-final class BreakMethodArgumentFixerTest extends AbstractSimpleFixerTestCase
+final class BreakMethodArgumentsFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
      * @dataProvider wrongToFixedCases()
@@ -30,7 +30,7 @@ final class BreakMethodArgumentFixerTest extends AbstractSimpleFixerTestCase
 
     protected function createFixer(): FixerInterface
     {
-        $fixer = new BreakMethodArgumentFixer();
+        $fixer = new BreakMethodArgumentsFixer();
         $fixer->setWhitespacesConfig(new WhitespacesFixerConfig());
 
         return $fixer;
